moved metadata for all other plots to bhom/wrapped and changed sunpath method to get the sunrise, sunset and noon times, azimuths and altitudes
