+++ LadybugTools_Engine/Python/src/ladybugtools_toolkit/bhom/wrapped/plot/heatmap.py
@@ -12,7 +12,8 @@ def heatmap(epw_file: str, data_type_key: str, colour_map: str, return_file: str
         from ladybug.epw import EPW
         from ladybug.datacollection import HourlyContinuousCollection
         from ladybugtools_toolkit.plot._heatmap import heatmap
-        from ladybugtools_toolkit.ladybug_extension.datacollection import collection_to_series, collection_metadata
+        from ladybugtools_toolkit.ladybug_extension.datacollection import collection_to_series
+        from ladybugtools_toolkit.bhom.wrapped.collection import collection_metadata
         from ladybugtools_toolkit.plot.utilities import figure_to_base64
         import matplotlib.pyplot as plt
