+++ LadybugTools_Engine/Python/src/ladybugtools_toolkit/bhom/wrapped/plot/sunpath.py
@@ -10,7 +10,7 @@ def sun_path(epw_file, analysis_period, size, return_file: str, save_path):
         from ladybug.epw import EPW, AnalysisPeriod
         from ladybug.datacollection import HourlyContinuousCollection
         from ladybug.sunpath import Sunpath
-        from ladybugtools_toolkit.ladybug_extension.sunpath import sunpath_metadata
+        from ladybugtools_toolkit.bhom.wrapped.sunpath_metadata import sunpath_metadata
         from ladybugtools_toolkit.plot.utilities import figure_to_base64
         import matplotlib.pyplot as plt
         from pathlib import Path
